requirements.txt
numpy==1.12.0
PyOpenGL==3.1.0
PyQt5==5.8.1
pyrr==0.8.2
